Receives phone number as a long and not an integer in the database's delete function

+++ MyDBHandler.java
@@ -47,7 +47,7 @@ public class MyDBHandler extends SQLiteOpenHelper {
     }
 
     // Delete a contact from the CONTACTS table
-    public void deleteContact(int contactphone) {
+    public void deleteContact(long contactphone) {
         SQLiteDatabase db = getWritableDatabase();
         db.execSQL("DELETE FROM " + TABLE_CONTACTS + " WHERE " + CONTACT_COLUMN_PHONE + " = " + contactphone + ";");
     }
